Update rail version via config not gemfile - Fixes thebeansgroup/januscli#21
