+++ lib/index.js
@@ -60,14 +60,6 @@ class Release extends Plugin {
 
   cliOptions() {
     return [
-      [
-        '-t, --test <n>',
-        'Test 2',
-        false,
-        function(val) {
-          console.log('arg called', arguments);
-        }
-      ]
     ]
   }
