+++ lib/version.js
@@ -1,7 +1,7 @@
 import semver from 'semver';
 import shell from 'shelljs';
 import NpmVersion from './version/npm';
-import GemVersion from './version/gem';
+import RailsVersion from './version/rails';
 
 /**
  * Semantic Version Class
@@ -35,37 +35,36 @@ class Version {
   */
 
   findReleaseType() {
-    if(['major', 'minor', 'patch'].indexOf(this.level) !== -1) {
+    if (['major', 'minor', 'patch'].indexOf(this.level) !== -1) {
       return this.level;
     }
 
     let labels = [];
 
-    this.prs.forEach(function(pr) {
-      pr.issues.forEach(function(issue) {
+    this.prs.forEach((pr) => {
+      pr.issues.forEach((issue) => {
         labels = labels.concat(issue.labels);
       });
     });
 
     // TODO: When do we major release?
-    if(labels.indexOf('enhancement') !== -1) {
+    if (labels.indexOf('enhancement') !== -1) {
       return 'minor';
-    }
-    else if(labels.indexOf('bug') !== -1) {
+    } else if (labels.indexOf('bug') !== -1) {
       return 'patch';
-    } else {
-      this.janus.error(
-        'Release:',
-        'Version can not be identified from issue labels.'
-      );
     }
+
+    this.janus.error(
+      'Release:',
+      'Version can not be identified from issue labels.'
+    );
   }
 
 
   getCurrentVersion() {
     const cmd = 'git describe --tags --abbrev=0';
-    return semver.clean(shell.exec(cmd, {silent: !this.janus.debug}).stdout);
-  } 
+    return semver.clean(shell.exec(cmd, { silent: !this.janus.debug }).stdout);
+  }
 
 
  /**
@@ -77,8 +76,8 @@ class Version {
   updateVersionFiles(version) {
     const npmVersion = new NpmVersion(this.janus);
     npmVersion.write(version);
-    const gemVersion = new GemVersion(this.janus);
-    gemVersion.write(version);
+    const railsVersion = new RailsVersion(this.janus);
+    railsVersion.write(version);
   }
 
 }
